Test exporting sampler (currently failing)
